Checking the Internet connection is made in a separate class
